MAINT: Split momentum balance and contact mechanics models
